requirements.txt
uvicorn==0.23.2
django==4.2.4
djangorestframework==3.14.0
adrf==0.1.1 # need for using djangorestframework with async
drf-standardized-errors==0.12.5
whitenoise==6.5.0
django-cors-headers==4.2.0
django-cacheops==7.0.1
requests-async==0.6.2
python-dotenv==1.0.0